requirements.txt
fpdf==1.7.2
pydantic_core==2.33.2
pandas==2.2.3
pydantic==2.11.4
langchain_core==0.1.23
idna==3.10
pytz==2025.2
matplotlib==3.9.4
pillow==11.2.1
zipp==3.21.0
fonttools==4.57.0
yarl==1.20.0
python_docx==1.1.2
pip==25.1.1
anyio==4.9.0
requests==2.32.3
attrs==25.3.0
blinker==1.9.0
frozenlist==1.6.0
pyparsing==3.2.3
pinecone_plugin_interface==0.0.7
charset_normalizer==3.4.2
typing_inspect==0.9.0
langchain_text_splitters==0.0.2
itsdangerous==2.2.0
langsmith==0.0.87
tqdm==4.67.1
exceptiongroup==1.2.2
jsonpatch==1.33
click==8.1.8
Pypdf2
Pdf2image
typing_inspection==0.4.0
jsonpointer==3.0.0
tzdata==2025.2
langchain_community==0.0.20
numpy==1.26.4
typing_extensions==4.13.2
httpcore==1.0.9
packaging==23.2
langchain_pinecone==0.1.3
async_timeout==4.0.3
python_dotenv==1.1.0
pinecone_plugin_inference==1.1.0
dataclasses_json==0.6.7
werkzeug==3.1.3
kiwisolver==1.4.7
certifi==2025.4.26
openai==1.77.0
contourpy==1.3.0
regex==2024.11.6
h11==0.16.0
sniffio==1.3.1
langchain==0.0.354
python_dateutil==2.9.0.post0
tenacity==8.5.0
urllib3==2.4.0
pinecone_client==5.0.1
cycler==0.12.1
jinja2==3.1.6
PyMySQL==1.1.1
langchain_openai==0.0.2.post1
aiohttp==3.11.18
distro==1.9.0
aiohappyeyeballs==2.6.1
requests_toolbelt==1.0.0
flask==3.1.0
propcache==0.3.1
lxml==5.4.0
aiosignal==1.3.2
PyYAML==6.0.2
httpx==0.28.1
flask_cors==5.0.1
importlib_metadata==8.7.0
sqlalchemy==2.0.40
orjson==3.10.18
multidict==6.4.3
marshmallow==3.26.1
crewai==0.1.24
annotated_types==0.7.0
setuptools==58.0.4
mypy_extensions==1.1.0
six==1.17.0
tiktoken==0.5.2
jiter==0.9.0
importlib_resources==6.5.2
tavily_python==0.7.1
MarkupSafe==3.0.2
Crawl4ai
Chroma